EIP196 implementation
Summary:
This diff does two things:
* Move modexp precompiled contract to a separate crate to avoid additional dependency for the VM itself.
* Implements add and mul bn128 precompiled contracts using Parity's forked version of bn crate.

Note: Upstream review process is ongoing, so this may land here sooner than it does in SputnikVM repo.

Reviewers: sorpaas

Reviewed By: sorpaas

Subscribers: jenkins

Differential Revision: https://source.that.world/D44
