+++ Cargo.toml
@@ -23,18 +23,14 @@ etcommon-block = { version = "0.3", default-features = false, optional = true }
 secp256k1-plus = { version = "0.5.7", optional = true }
 libsecp256k1 = { version = "0.1", optional = true }
 
-num-bigint = { version = "0.1", optional = true }
-
 [dev-dependencies]
 etcommon-hexutil = "0.2"
 
 [features]
-default = ["std", "c-secp256k1", "precompiled-modexp"]
+default = ["std", "c-secp256k1"]
 c-secp256k1 = ["secp256k1-plus", "etcommon-block/c-secp256k1"]
 rust-secp256k1 = ["libsecp256k1", "etcommon-block/rust-secp256k1"]
 std = ["etcommon-block-core/std", "etcommon-rlp/std", "etcommon-bigint/std", "etcommon-block"]
-
-precompiled-modexp = ["num-bigint"]
 
 [workspace]
 members = [
@@ -43,4 +39,6 @@ members = [
   "./gethrpc",
   "./stateful",
   "./cli",
+  "./precompiled/modexp",
+  "./precompiled/bn128",
 ]
