+++ precompiled/modexp/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "sputnikvm-precompiled-modexp"
+version = "0.0.1"
+license = "Apache-2.0"
+authors = ["Wei Tang <hi@that.world>"]
+repository = "https://github.com/ethereumproject/sputnikvm"
+
+[dependencies]
+sputnikvm = { version = "0.9", path = "../.." }
+etcommon-bigint = { version = "0.2", default-features = false }
+num-bigint = "0.1"
